rename bigtest/slow_test
