+++ Makefile
@@ -106,8 +106,8 @@ test: build-gondar
 	${BUILD_DIR}/tests -platform offscreen
 
 # Run tests which take quite some time
-bigtest: build-gondar
-	${BUILD_DIR}/bigtests -platform offscreen
+slowtest: build-gondar
+	${BUILD_DIR}/slowtests -platform offscreen
 
 
 update-submodules:
